feat: limits the types to a smaler scope, and customizes descriptions (#13)
This commit removes some commit types to make it easier to use for developers. The descriptions are
updated to include more relevant and modern scenarios. Like exchanging gulp for webpack as an
example of a bundling-/build-tool.

BREAKING CHANGE: Some types are not included any more: style, perf. refactor can be used instead

Co-authored-by: Julius Celik <julius.celik@digitalroute.com>